refactored SLP_Ade project folder: split config into config.py, renamed multi_simulations.py to simulation.py, adjusted all references
