+++ SLP_Ade/simulation.py
@@ -1,7 +1,6 @@
-"""
-Multi-run LPG example template
+"""Multi-run LPG simulation logic.
 
-What this script does
+What this module does
 - Runs multiple LPG simulations for each household template selected by
     `HOUSEHOLD_TEMPLATE_KEYS`.
 - For each template it iterates over `CLIMATE_SET_KEYS` and
@@ -10,24 +9,16 @@ What this script does
   `RUNS_PER_COMBO` independent simulations with different random seeds to
   inspect stochastic variability.
 
+All tunable parameters live in `config.py`. This module holds the shared
+execution primitive `run_lpg_simulation()` plus the sequential (non-SLURM)
+runner. The SLURM worker `run_task.py` also calls `run_lpg_simulation()`, so
+both execution paths share the exact same primitive.
+
 Generic collector idea
 - `collect_lpg_members(...)` can collect all predefined members of a given
     type from any LPG static class via introspection (inspect).
 - This is used for templates, geographic locations, temperature profiles and
     transport sets, so lists do not need to be manually copied from `lpgdata`.
-
-Configuration
-- `HOUSEHOLD_TEMPLATE_KEYS`: template names from `lpgdata.HouseholdTemplates`.
-  Set to `None` to run all available templates.
-- `CLIMATE_SET_KEYS`: tuples of
-    `(geographic_location_key, temperature_profile_key, tag)`.
-    This keeps LPG's geographic location and weather profile separate, while
-    still letting you define meaningful paired presets. Set to `None` to run
-    all possible geographic-location/temperature-profile combinations.
-- `TRANSPORT_VARIANT_KEYS`: `TransportVariantKey` entries with key names for
-    LPG transport sets. Keys are resolved against the corresponding
-    `lpgdata.*Sets` classes.
-- `RUNS_PER_COMBO`: number of different seeds per parameter combination.
 
 Outputs
 - Per-run CSV files are written to `multi_runs_output` (if SAVE_CSV=True).
@@ -37,9 +28,10 @@ Outputs
 - `runs_metadata.csv` summarizes all successful runs.
 
 Run
-    python examples/multi_simulations.py
+    python SLP_Ade/simulation.py
 """
 
+import sys
 from pathlib import Path
 from dataclasses import dataclass
 import glob
@@ -50,49 +42,46 @@ from typing import Any, Iterable, Optional
 
 import pandas as pd
 
+# Make the repo root importable so `from SLP_Ade.config import ...` works both
+# when this module is imported as part of the package and when it is run
+# directly as a script.
+_REPO_ROOT = Path(__file__).resolve().parents[1]
+if str(_REPO_ROOT) not in sys.path:
+    sys.path.insert(0, str(_REPO_ROOT))
+
 from pylpg import lpg_execution, lpgdata
 from pylpg.lpgpythonbindings import EnergyIntensityType, JsonReference
 
-
-OUTPUT_DIR = Path("multi_runs_output")
-OUTPUT_DIR.mkdir(exist_ok=True)
-
-
-# Output format options
-SAVE_CSV = False  # Save individual CSV files per run
-SAVE_HDF5 = True  # Save runs to HDF5 files (one file per household template)
-
-
-def get_attr_key(cls: type, attr_value: Any) -> str:
-    """Get the attribute name from a class for a given attribute value.
-    
-    :param type cls: The class to search for the attribute.
-    :param Any attr_value: The attribute value to find the name for.
-    :return str: The attribute name.
-    :raises ValueError: If the attribute is not found in the class.
-    """
-    for name, value in inspect.getmembers(cls):
-        if not name.startswith("_") and value is attr_value:
-            return name
-    raise ValueError(f"Attribute not found in {cls.__name__}")
-
+from SLP_Ade.config import (
+    CLIMATE_SET_KEYS,
+    HOUSEHOLD_TEMPLATE_KEYS,
+    HOUSETYPE,
+    LPG_BINARY_PATH,
+    OUTPUT_DIR,
+    SAVE_CSV,
+    SAVE_HDF5,
+    TRANSPORT_VARIANT_KEYS,
+    TransportVariantKey,
+    YEAR,
+    get_runs_for_combo,
+)
 
 
 def prompt_clean_output_dir() -> None:
     """Ask user if they want to delete existing output files.
-    
+
     Prompts the user interactively to delete CSV and HDF5 files in OUTPUT_DIR.
     Lists existing files and waits for yes/no confirmation.
-    
+
     :return None: No return value.
     """
     csv_files = glob.glob(str(OUTPUT_DIR / "*.csv"))
     hdf5_files = glob.glob(str(OUTPUT_DIR / "*.h5"))
-    
+
     if not csv_files and not hdf5_files:
         print(f"Output directory '{OUTPUT_DIR}' is empty. Ready to start.")
         return
-    
+
     print(f"\nFound existing output files in '{OUTPUT_DIR}':")
     if csv_files:
         print(f"  - {len(csv_files)} CSV files")
@@ -104,10 +93,10 @@ def prompt_clean_output_dir() -> None:
         print(f"  - {len(hdf5_files)} HDF5 files")
         for f in sorted(hdf5_files):
             print(f"    - {Path(f).name}")
-    
+
     while True:
         response = input("\nDelete all existing multirun output files for this run? (yes/no): ").strip().lower()
-        
+
         if response in ("yes", "y"):
             for f in csv_files:
                 Path(f).unlink()
@@ -121,9 +110,10 @@ def prompt_clean_output_dir() -> None:
         else:
             print("Invalid response. Please enter 'yes/y' or 'no/n'.")
 
+
 def safe_name(s: str) -> str:
     """Convert string to filesystem-safe name by replacing/removing special characters.
-    
+
     :param str s: The string to convert.
     :return str: Filesystem-safe version of the string.
     """
@@ -133,7 +123,7 @@ def safe_name(s: str) -> str:
 
 def split_dataframe_by_type(df: pd.DataFrame) -> dict[str, pd.DataFrame]:
     """Split DataFrame columns by data type prefix (e.g., 'Electricity_HH1' -> 'Electricity').
-    
+
     :param pd.DataFrame df: The DataFrame to split.
     :return dict[str, pd.DataFrame]: Dictionary mapping data type to DataFrame.
     """
@@ -148,7 +138,7 @@ def split_dataframe_by_type(df: pd.DataFrame) -> dict[str, pd.DataFrame]:
 
 def collect_lpg_members(container: Any, expected_type: type) -> dict[str, Any]:
     """Collect public class members of `container` that match `expected_type`.
-    
+
     :param Any container: The class or object to collect members from.
     :param type expected_type: The type to filter members by.
     :return dict[str, Any]: Dictionary mapping member names to values.
@@ -166,7 +156,7 @@ def select_by_keys(
     label: str,
 ) -> list[Any]:
     """Select values from `available` by key list, or all values if keys is None.
-    
+
     :param dict[str, Any] available: Dictionary of available values.
     :param Optional[Iterable[str]] keys: Keys to select, or None for all values.
     :param str label: Label for error messages.
@@ -190,7 +180,7 @@ def resolve_optional_key(
     label: str,
 ) -> Optional[JsonReference]:
     """Resolve an optional key to a JsonReference.
-    
+
     :param dict[str, JsonReference] available: Dictionary of available references.
     :param Optional[str] key: The key to resolve, or None.
     :param str label: Label for error messages.
@@ -210,7 +200,7 @@ def make_climate_variants(
     climate_keys: Optional[list[tuple[str, Optional[str], str]]],
 ) -> list[tuple[JsonReference, Optional[JsonReference], str]]:
     """Create climate variant combinations.
-    
+
     :param dict[str, JsonReference] all_geographic_locations: Available geographic locations.
     :param dict[str, JsonReference] all_temperature_profiles: Available temperature profiles.
     :param Optional[list[tuple[str, Optional[str], str]]] climate_keys: List of (location_key, temp_key, tag) tuples, or None for all combinations.
@@ -243,27 +233,9 @@ def make_climate_variants(
 
 
 @dataclass(frozen=True)
-class TransportVariantKey:
-    """Configuration key for a transport variant.
-    
-    Attributes:
-        simulate_transportation: Whether to enable transportation simulation.
-        charging_set_key: Key for charging station set (or None).
-        transport_device_set_key: Key for transportation device set (or None).
-        travel_route_set_key: Key for travel route set (or None).
-        tag: Human-readable tag for this variant.
-    """
-    simulate_transportation: bool
-    charging_set_key: Optional[str]
-    transport_device_set_key: Optional[str]
-    travel_route_set_key: Optional[str]
-    tag: str
-
-
-@dataclass(frozen=True)
 class TransportVariant:
     """Resolved transport variant with JsonReferences.
-    
+
     Attributes:
         simulate_transportation: Whether to enable transportation simulation.
         charging_set: Charging station set reference (or None).
@@ -285,7 +257,7 @@ def make_transport_variants(
     variant_keys: list[TransportVariantKey],
 ) -> list[TransportVariant]:
     """Create transport variant combinations from variant keys.
-    
+
     :param dict[str, JsonReference] all_charging_sets: Available charging station sets.
     :param dict[str, JsonReference] all_transport_device_sets: Available transportation device sets.
     :param dict[str, JsonReference] all_travel_route_sets: Available travel route sets.
@@ -314,83 +286,12 @@ def make_transport_variants(
     ]
 
 
-# ---- CONFIG ------------------------------------------------------------------------------------------------------------------------------------------
-YEAR = 2022
-
-# Set to None to use all templates in lpgdata.HouseholdTemplates.
-HOUSEHOLD_TEMPLATE_KEYS = None  # [
-    # get_attr_key(lpgdata.HouseholdTemplates, lpgdata.HouseholdTemplates.CHR01_Couple_both_at_Work),
-    # get_attr_key(lpgdata.HouseholdTemplates, lpgdata.HouseholdTemplates.CHR03_Family_1_child_both_at_work),
-# ]
-
-# Climate presets keep geographic location and temperature profile separate.
-# (geographic_location_key, temperature_profile_key, tag)
-CLIMATE_SET_KEYS = [
-    (
-        get_attr_key(lpgdata.GeographicLocations, lpgdata.GeographicLocations.Germany_Berlin),
-        get_attr_key(lpgdata.TemperatureProfiles, lpgdata.TemperatureProfiles.Berlin_Germany_1996_from_Deutscher_Wetterdienst_DWD_www_dwd_de),
-        "berlin_loc_berlin_temp",
-    ),
-    (
-        get_attr_key(lpgdata.GeographicLocations, lpgdata.GeographicLocations.Germany_Hamburg),
-        get_attr_key(lpgdata.TemperatureProfiles, lpgdata.TemperatureProfiles.Hamburg_Germany_2007_from_Deutscher_Wetterdienst_DWD_www_dwd_de),
-        "hamburg_loc_hamburg_temp",
-    ),
-    (
-        get_attr_key(lpgdata.GeographicLocations, lpgdata.GeographicLocations.Germany_Chemnitz),
-        get_attr_key(lpgdata.TemperatureProfiles, lpgdata.TemperatureProfiles.Dresden_Germany_2000_from_Deutscher_Wetterdienst_DWD_www_dwd_de),
-        "chemnitz_loc_dresden_temp",
-    ),
-]
-# Set to None to generate all location/temperature-profile combinations.
-
-# Key-based transport presets.
-TRANSPORT_VARIANT_KEYS = [
-    TransportVariantKey(False, None, None, None, "no_transport"),
-    TransportVariantKey(
-        True,
-        get_attr_key(lpgdata.ChargingStationSets, lpgdata.ChargingStationSets.Charging_At_Home_with_03_7_kW_output_results_to_Car_Electricity),
-        get_attr_key(lpgdata.TransportationDeviceSets, lpgdata.TransportationDeviceSets.Bus_and_two_30_km_h_Cars),
-        get_attr_key(lpgdata.TravelRouteSets, lpgdata.TravelRouteSets.Travel_Route_Set_for_30km_Commuting_Distance),
-        "home_charge_bus_cars_30km",
-    ),
-]
-
-HOUSETYPE = lpgdata.HouseTypes.HT20_Single_Family_House_no_heating_cooling
-
-#: Custom binary path for LPG. Set to None to use the official release downloaded automatically by the package.
-LPG_BINARY_PATH = None 
-
-# Define runs per combination. You can specify:
-# - A dict mapping combo_tag patterns to run counts
-# - Or use a function to determine runs based on parameters
-RUNS_PER_COMBO_MAP = {
-    "no_transport": 1,           # Baseline: 1 run only
-    "home_charge_bus_cars_30km": 3,  # Transport variants: 3 runs
-}
-
-def get_runs_for_combo(combo_tag: str) -> int:
-    """Determine the number of runs for a given combination tag.
-    
-    Looks up the combo_tag in RUNS_PER_COMBO_MAP to find matching patterns.
-    Returns the configured number of runs, or a default of 2 if no pattern matches.
-    
-    :param str combo_tag: The combination tag to look up.
-    :return int: Number of runs for this combination.
-    """
-    for pattern, runs in RUNS_PER_COMBO_MAP.items():
-        if pattern in combo_tag:
-            return runs
-    return 2  # Default fallback: 2 runs for unmapped combinations
-
-# ---- END CONFIG ------------------------------------------------------------------------------------------------------------------------------------------
-
 def _print_lpg_binary_source() -> None:
     """Print the source of the LPG binary being used.
-    
+
     :return None: No return value.
     """
-    if LPG_BINARY_PATH is None: 
+    if LPG_BINARY_PATH is None:
         print("LPG binary source: official release downloaded automatically.")
     else:
         print(f"LPG binary source: custom binary path {LPG_BINARY_PATH}")
@@ -408,10 +309,10 @@ def save_as_HDF5(
     data_types: dict[str, pd.DataFrame],
 ) -> None:
     """Save simulation results to HDF5 file with hierarchical structure.
-    
+
     Creates or appends to an HDF5 file named after the template. Data is organized as:
     /climate_tag/transport_tag/run_N/data_type
-    
+
     :param str tmpl_name: Template name for the HDF5 filename.
     :param JsonReference geographic_location: Geographic location reference.
     :param Optional[JsonReference] temperature_profile: Temperature profile reference.
@@ -448,10 +349,10 @@ def save_as_HDF5(
 
 def create_combo_tag(tmpl_name: str, climate_name: str, transport_tag: str) -> str:
     """Create a combination tag from template, climate, and transport names.
-    
+
     Combines the three components with double underscores, using safe_name()
     to ensure filesystem compatibility.
-    
+
     :param str tmpl_name: Template name.
     :param str climate_name: Climate variant name.
     :param str transport_tag: Transport variant tag.
@@ -470,7 +371,7 @@ def run_lpg_simulation(
     """Build a household and execute one LPG simulation.
 
     This is the core execution primitive shared between interactive
-    (multi_simulations) and SLURM-array (run_task) modes.
+    (simulation) and SLURM-array (run_task) modes.
 
     :param str tmpl: Template name string (value of a HouseholdTemplates attribute).
     :param TransportVariant transport_variant: Resolved transport configuration.
@@ -525,7 +426,7 @@ def execute_single_run(
     num_runs: int,
 ) -> Optional[dict[str, Any]]:
     """Execute a single LPG simulation run.
-    
+
     :param str tmpl: Template identifier.
     :param str tmpl_name: Template name for output files.
     :param JsonReference geographic_location: Geographic location reference.
@@ -557,17 +458,17 @@ def execute_single_run(
             return None
 
         filename_base = f"{combo_tag}__seed{seed}__run{run_idx + 1}"
-        
+
         # Split dataframe by data type
         data_types = split_dataframe_by_type(df)
-        
+
         # Save to CSV if enabled
         if SAVE_CSV:
             for data_type, type_df in data_types.items():
                 out_csv = OUTPUT_DIR / (safe_name(f"{filename_base}__{data_type}") + ".csv")
                 type_df.to_csv(out_csv)
             print(f"  Saved {len(data_types)} data types to CSV: {', '.join(sorted(data_types.keys()))}")
-        
+
         # Save to HDF5 if enabled (one file per household template)
         if SAVE_HDF5:
             save_as_HDF5(
@@ -609,10 +510,10 @@ def execute_all_runs(
     transport_variants: list[TransportVariant],
 ) -> tuple[list[dict[str, Any]], int]:
     """Execute all simulation runs for given parameter combinations.
-    
+
     Iterates through all combinations of templates, climate variants, and transport
     variants, executing multiple runs per combination based on configuration.
-    
+
     :param list[str] household_templates: List of household template names.
     :param list[tuple[JsonReference, Optional[JsonReference], str]] climate_sets: List of (location, temp_profile, tag) tuples.
     :param list[TransportVariant] transport_variants: List of transport variant configurations.
@@ -626,14 +527,14 @@ def execute_all_runs(
         if tmpl is None:
             raise ValueError("Household template is None")
         tmpl_name = tmpl
-        
+
         for geographic_location, temperature_profile, climate_tag in climate_sets:
             climate_name = climate_tag
-            
+
             for transport_variant in transport_variants:
                 combo_tag = create_combo_tag(tmpl_name, climate_name, transport_variant.tag)
                 num_runs = get_runs_for_combo(combo_tag)
-                
+
                 # Execute multiple runs with different seeds for this combination
                 for run_idx in range(num_runs):
                     metadata = execute_single_run(
@@ -648,17 +549,17 @@ def execute_all_runs(
                         run_idx,
                         num_runs,
                     )
-                    
+
                     if metadata is not None:
                         meta_rows.append(metadata)
                         total += 1
-    
+
     return meta_rows, total
 
 
 def print_summary(meta_df: pd.DataFrame, total: int) -> None:
     """Print summary of completed simulation runs.
-    
+
     :param pd.DataFrame meta_df: DataFrame containing metadata for all runs.
     :param int total: Total number of successful runs.
     :return None: No return value.
@@ -669,13 +570,13 @@ def print_summary(meta_df: pd.DataFrame, total: int) -> None:
     if SAVE_HDF5:
         hdf5_files = meta_df['hdf5_file'].dropna().unique()
         output_summary.append(f"{len(hdf5_files)} HDF5 file(s) (one per household template)")
-    
+
     print(
         f"\nFinished {total} successful runs.\n"
         f"Output: {' and '.join(output_summary)}\n"
         f"Metadata: {OUTPUT_DIR / 'runs_metadata.csv'}"
     )
-    
+
     if SAVE_HDF5:
         hdf5_files = meta_df['hdf5_file'].dropna().unique()
         print(f"\nHDF5 files created:")
@@ -686,14 +587,14 @@ def print_summary(meta_df: pd.DataFrame, total: int) -> None:
 
 def run_all() -> None:
     """Run all configured LPG simulations and save results.
-    
+
     Main execution function that:
     1. Collects all available LPG members (templates, locations, etc.)
     2. Creates parameter combinations from configuration
     3. Runs simulations for each combination with multiple seeds
     4. Saves results to CSV and/or HDF5 files
     5. Generates metadata CSV summarizing all runs
-    
+
     :return None: No return value.
     """
     _print_lpg_binary_source()
@@ -740,7 +641,6 @@ def run_all() -> None:
     meta_df = pd.DataFrame(meta_rows)
     meta_df.to_csv(OUTPUT_DIR / "runs_metadata.csv", index=False)
     print_summary(meta_df, total)
-
 
 
 if __name__ == "__main__":
